Fixed NS.class bug
The class file was larger than max short size (that is my guess) and it was not reading the entire file as bytes in from the zip file because I was only calling one read. Now I am reading it one byte at a time (inefficient but it works for now and class files are small) and it works correctly. Yay! Time to work on actual hooks now!