Can now find nearest NPC and perform actions
Most of the NPC interaction is done. It can now find the nearest NPC by id and perform a given action on the NPC. The command ::ba works for banking (at least in Falador because those are the NPC ids I added). The only thing that doesn't work is the nearest part because the getX() and getY() may be the wrong values (or need to be transformed somehow) because it isn't going to the nearest NPC always.