Restructured code into multiple projects and made many improvements
API is a separate Maven project which can be set as a dependency for a bot script project (see TestBotScript). I made a test bot script and compiled it as a JAR. It can be loaded in-game via ::load jar_path main_script_class

It's a lot easier being able to use the IDE to write bot scripts. This is feeling like a normal fully-working bot already because you can write your own scripts pretty easily